Finish order section, organize buttons
