+++ frontend/src/css_global/0_abstracts/_mixins.scss
@@ -1,14 +1,19 @@
-@use './variables';
+@use "./variables";
 
 @mixin size($width, $height: $width) {
     width: $width;
     height: $height;
 }
 
-@mixin button-hover($color, $b-radius) {
-  background-color: $color;
-  border-radius: $b-radius;
-  transition: background-color 0.5s, border-radius 0.5s;
+@mixin horizontal-align($percent-x, $percent-y: 0) {
+    left: $percent-x;
+    transform: translate(-$percent-x, -$percent-y);
+}
+
+@mixin align($percent-x, $percent-y: 0) {
+    left: $percent-x;
+    top: $percent-y;
+    transform: translate(-$percent-x, -$percent-y);
 }
 
 /// Responsive breakpoint manager
@@ -17,20 +22,16 @@
 /// @requires $breakpoints
 // Taken from https://sass-guidelin.es/#responsive-web-design-and-breakpoints
 @mixin respond-to($breakpoint) {
-  $raw-query: map-get(variables.$breakpoints, $breakpoint);
+    $raw-query: map-get(variables.$breakpoints, $breakpoint);
 
-  @if $raw-query {
-    $query: if(
-      type-of($raw-query) == 'string',
-      unquote($raw-query),
-      inspect($raw-query)
-    );
+    @if $raw-query {
+        $query: if(type-of($raw-query) == "string", unquote($raw-query), inspect($raw-query));
 
-    @media #{$query} {
-      @content;
-    }
-  } @else {
-    @error 'No value found for `#{$breakpoint}`. '
+        @media #{$query} {
+            @content;
+        }
+    } @else {
+        @error 'No value found for `#{$breakpoint}`. '
          + 'Please make sure it is defined in `$breakpoints` map.';
-  }
+    }
 }
