+++ frontend/src/css_global/1_base/_reset.scss
@@ -24,6 +24,11 @@ time, mark, audio, video {
 	font-family: Arial, Helvetica, sans-serif;
 	vertical-align: baseline;
 }
+
+html, body {
+	height: 100%;
+}
+
 /* HTML5 display-role reset for older browsers */
 article, aside, details, figcaption, figure, 
 footer, header, hgroup, menu, nav, section {
