+++ frontend/src/css_global/2_layouts/_background.scss
@@ -1,8 +1,19 @@
 @use '../0_abstracts/' as *;
 
 .header {
+    height: 100vh; // Fallback for older browsers
+    @include size(100%, 100%);
+}
+
+.background {
     background-image: linear-gradient(rgba(0,0,0,0.6), rgba(0,0,0,0.6)), url(../../assets/img/mainpizza.jpg);
     @include size(100%, 100vh);
     background-size: cover;
-    background-attachment: fixed;
+    position: fixed;
+    left: 0;
+    top: 0;
+}
+
+#root {
+    height: 100%;
 }
