+++ frontend/src/layouts/Navbar.js
@@ -1,7 +1,7 @@
 import { useState } from 'react'
 
 const Navbar = () => {
-    const logo = require('../assets/img/velo_city_logo.png');
+    const logo = require('assets/img/velo_city_logo.png');
     const [expandMenu, setExpandMenu] = useState(false)
 
     function expandMobileMenu() {
@@ -9,18 +9,18 @@ const Navbar = () => {
     }
 
     return (
-        <nav className="navbar">
-            <a className="nav-logo" href="/"><img src={logo} /></a>
-            <a className ="nav-name" href="/">Velo City Pizza</a>
-            <div className="mobile-burger" onClick={expandMobileMenu}>
-                <span className="bar"></span>
-                <span className="bar"></span>
-                <span className="bar"></span>
+        <nav className='navbar'>
+            <a className='nav-logo' href='/'><img src={logo} alt='Velo City Logo'/></a>
+            <a className ='nav-name' href='/'>Velo City Pizza</a>
+            <div className='mobile-burger' onClick={expandMobileMenu}>
+                <span className='bar'></span>
+                <span className='bar'></span>
+                <span className='bar'></span>
             </div>
             <ul className={`nav-links${expandMenu ? ' active' : ''}`}>
-                <li><a href="/">Home</a></li>
-                <li><a href="/order_ahead">Order</a></li>
-                <li><a href="/menu">Menu</a></li>
+                <li><a href='/'>Home</a></li>
+                <li><a href='/order_ahead'>Order</a></li>
+                <li><a href='/menu'>Menu</a></li>
             </ul>
             
         </nav>
